@@ -1,88 +1,89 @@
 {
   "name": "juicy",
-  "version": "1.0.0",
+  "version": "1.0.3",
   "description": "A Vue.js admin",
   "author": "Pan <panfree23@gmail.com>",
+  "license": "MIT",
   "private": true,
   "scripts": {
     "dev": "node build/dev-server.js",
-    "build:prod": "NODE_ENV=production node build/build.js",
-    "build:sit": "NODE_ENV=sit node build/build.js",
-    "build:sit-preview": "NODE_ENV=sit npm_config_preview=true  npm_config_report=true node build/build.js",
+    "build:prod": "cross-env NODE_ENV=production node build/build.js",
+    "build:sit": "cross-env NODE_ENV=sit node build/build.js",
+    "build:sit-preview": "cross-env NODE_ENV=sit npm_config_preview=true  npm_config_report=true node build/build.js",
     "lint": "eslint --ext .js,.vue src"
   },
   "dependencies": {
-    "axios": "0.15.3",
-    "codemirror": "5.22.0",
-    "dropzone": "4.3.0",
-    "echarts": "3.4.0",
-    "element-ui": "1.2.9",
+    "axios": "0.16.2",
+    "codemirror": "5.26.0",
+    "dropzone": "5.1.0",
+    "echarts": "3.6.1",
+    "element-ui": "1.3.6",
     "file-saver": "1.3.3",
     "jquery": "3.1.1",
-    "js-cookie": "2.1.3",
+    "js-cookie": "2.1.4",
     "jsonlint": "1.6.2",
     "mockjs": "1.0.1-beta3",
     "normalize.css": "3.0.2",
     "nprogress": "0.2.0",
-    "showdown": "1.6.4",
+    "screenfull": "3.2.2",
+    "showdown": "1.7.1",
     "simplemde": "1.11.2",
-    "vue": "2.2.6",
-    "vue-multiselect": "2.0.0-beta.14",
-    "vue-router": "2.3.0",
-    "vuedraggable": "2.8.4",
-    "vuex": "2.2.1",
-    "xlsx": "0.8.1",
-    "vue-count-to":"1.0.5"
+    "sortablejs": "1.5.1",
+    "vue": "2.3.3",
+    "vue-count-to": "1.0.5",
+    "vue-multiselect": "2.0.0-beta.15",
+    "vue-router": "2.5.3",
+    "vuedraggable": "2.13.1",
+    "vuex": "2.3.1",
+    "xlsx": "0.8.1"
   },
   "devDependencies": {
-    "autoprefixer": "6.7.2",
-    "axios-mock-adapter": "^1.8.1",
-    "babel-core": "6.22.1",
-    "babel-eslint": "7.1.1",
-    "babel-loader": "6.2.10",
-    "babel-plugin-transform-runtime": "6.22.0",
-    "babel-preset-es2015": "6.22.0",
-    "babel-preset-stage-2": "6.22.0",
-    "babel-register": "6.22.0",
+    "autoprefixer": "7.1.1",
+    "babel-core": "6.25.0",
+    "babel-eslint": "7.2.3",
+    "babel-loader": "7.0.0",
+    "babel-plugin-transform-runtime": "6.23.0",
+     "babel-preset-env": "1.5.2",
+    "babel-preset-stage-2": "6.24.1",
+    "babel-register": "6.24.1",
     "chalk": "1.1.3",
     "connect-history-api-fallback": "1.3.0",
     "copy-webpack-plugin": "4.0.1",
-    "css-loader": "0.26.1",
-    "eslint": "3.14.1",
-    "eslint-config-airbnb-base": "11.0.1",
-    "eslint-friendly-formatter": "2.0.7",
+    "cross-env": "5.0.1",
+    "css-loader": "0.28.4",
+    "eslint": "3.19.0",
+    "eslint-friendly-formatter": "3.0.0",
     "eslint-import-resolver-webpack": "0.8.1",
-    "eslint-loader": "1.6.3",
-    "eslint-plugin-html": "2.0.0",
-    "eslint-plugin-import": "2.2.0",
+    "eslint-loader": "1.7.1",
+    "eslint-plugin-html": "3.0.0",
+    "eslint-plugin-import": "2.3.0",
     "eventsource-polyfill": "0.9.6",
-    "express": "4.14.1",
-    "extract-text-webpack-plugin": "2.0.0",
-    "file-loader": "0.10.0",
-    "friendly-errors-webpack-plugin": "1.1.3",
+    "express": "4.15.3",
+    "extract-text-webpack-plugin": "2.1.2",
+    "file-loader": "0.11.2",
+    "friendly-errors-webpack-plugin": "1.6.1",
     "function-bind": "1.1.0",
     "html-webpack-plugin": "2.28.0",
-    "http-proxy-middleware": "0.17.3",
-    "node-sass": "3.7.0",
+    "http-proxy-middleware": "0.17.4",
+    "node-sass": "^4.5.0",
     "opn": "4.0.2",
     "optimize-css-assets-webpack-plugin": "1.3.0",
     "ora": "1.1.0",
     "pushstate-server": "2.1.0",
     "rimraf": "2.6.0",
-    "sass-loader": "4.0.2",
+    "sass-loader": "6.0.5",
     "script-loader": "0.7.0",
     "semver": "5.3.0",
-    "style-loader": "0.13.1",
-    "url-loader": "0.5.7",
-    "vue-loader": "11.3.4",
-    "vue-style-loader": "2.0.0",
-    "vue-template-compiler": "2.2.6",
-    "webpack": "2.2.1",
-    "webpack-bundle-analyzer": "2.2.1",
-    "webpack-dashboard": "0.2.1",
-    "webpack-dev-middleware": "1.10.0",
-    "webpack-hot-middleware": "2.16.1",
-    "webpack-merge": "2.6.1"
+    "style-loader": "0.17.0",
+    "url-loader": "0.5.8",
+    "vue-loader": "12.2.1",
+    "vue-style-loader": "3.0.1",
+    "vue-template-compiler": "2.3.3",
+    "webpack": "2.6.1",
+    "webpack-bundle-analyzer": "2.8.2",
+    "webpack-dev-middleware": "1.10.2",
+    "webpack-hot-middleware": "2.18.0",
+    "webpack-merge": "4.1.0"
   },
   "engines": {
     "node": ">= 4.0.0",
