@@ -1,28 +1,8 @@
-import fetch, { tpFetch } from 'utils/fetch';
+import fetch from 'utils/fetch';
 
 export function getToken() {
   return fetch({
-    url: '/qiniu/upload/token',
+    url: '/qiniu/upload/token', // 假地址 自行替换
     method: 'get'
-  });
-}
-export function upload(data) {
-  return tpFetch({
-    url: 'https://upload.qbox.me',
-    method: 'post',
-    data
-  });
-}
-
-
-/* 外部uri转七牛uri*/
-export function netUpload(token, net_url) {
-  const imgData = {
-    net_url
-  };
-  return fetch({
-    url: '/qiniu/upload/net/async',
-    method: 'post',
-    data: imgData
   });
 }
