@@ -105,8 +105,7 @@
                   return idx * 20;
                 }
               }]
-            }
-                )
+            })
         }
       }
     }
