@@ -1,26 +1,25 @@
-import axios from 'axios';
-import MockAdapter from 'axios-mock-adapter';
+import Mock from 'mockjs';
 import loginAPI from './login';
 import articleAPI from './article';
 import article_tableAPI from './article_table';
 import remoteSearchAPI from './remoteSearch';
-const mock = new MockAdapter(axios);
+
 
 // 登录相关
-mock.onPost('/login/loginbyemail').reply(loginAPI.loginByEmail);
-mock.onPost('/login/logout').reply(loginAPI.logout);
-mock.onGet('/user/info').reply(loginAPI.getInfo);
+Mock.mock(/\/login\/loginbyemail/, 'post', loginAPI.loginByEmail);
+Mock.mock(/\/login\/logout/, 'post', loginAPI.logout);
+Mock.mock(/\/user\/info\.*/, 'get', loginAPI.getInfo)
 
-// 文章相关
-mock.onGet('/article/list').reply(articleAPI.getList);
-mock.onGet('/article/detail').reply(articleAPI.getArticle);
+// // 文章相关
+Mock.mock(/\/article\/list/, 'get', articleAPI.getList);
+Mock.mock(/\/article\/detail/, 'get', articleAPI.getArticle);
 
-// table example相关
-mock.onGet('/article_table/list').reply(article_tableAPI.getList);
-mock.onGet('/article_table/pv').reply(article_tableAPI.getPv);
+// // table example相关
+Mock.mock(/\/article_table\/list/, 'get', article_tableAPI.getList);
+Mock.mock(/\/article_table\/p/, 'get', article_tableAPI.getPv);
 
-// 搜索相关
-mock.onGet('/search/user').reply(remoteSearchAPI.searchUser);
+// // 搜索相关
+Mock.mock(/\/search\/user/, 'get', remoteSearchAPI.searchUser);
 
 
-export default mock;
+export default Mock;
