@@ -1,60 +1,45 @@
+import { param2Obj } from 'utils';
+
 const userMap = {
   admin: {
     role: ['admin'],
     token: 'admin',
     introduction: '我是超级管理员',
-    avatar: 'https://wdl.wallstreetcn.com/48a3e1e0-ea2c-4a4e-9928-247645e3428b',
-    name: '超级管理员小潘'
+    avatar: 'https://wpimg.wallstcn.com/f778738c-e4f8-4870-b634-56703b4acafe.gif',
+    name: '超级管理员小潘',
+    uid: '001'
   },
   editor: {
     role: ['editor'],
     token: 'editor',
     introduction: '我是编辑',
-    avatar: 'https://wdl.wallstreetcn.com/48a3e1e0-ea2c-4a4e-9928-247645e3428b',
-    name: '普通编辑小张'
+    avatar: 'https://wpimg.wallstcn.com/f778738c-e4f8-4870-b634-56703b4acafe.gif',
+    name: '普通编辑小张',
+    uid: '002'
 
   },
   developer: {
     role: ['develop'],
     token: 'develop',
     introduction: '我是开发',
-    avatar: 'https://wdl.wallstreetcn.com/48a3e1e0-ea2c-4a4e-9928-247645e3428b',
-    name: '工程师小王'
+    avatar: 'https://wpimg.wallstcn.com/f778738c-e4f8-4870-b634-56703b4acafe.gif',
+    name: '工程师小王',
+    uid: '003'
   }
 }
 
 export default {
   loginByEmail: config => {
-    const { email } = config.params;
-    return new Promise((resolve, reject) => {
-      if (userMap[email.split('@')[0]]) {
-        setTimeout(() => {
-          resolve([200, {
-            data: userMap[email.split('@')[0]]
-          }]);
-        }, 500);
-      } else {
-        reject('账号不正确')
-      }
-    })
+    const { email } = JSON.parse(config.body);
+    return userMap[email.split('@')[0]];
   },
   getInfo: config => {
-    const { token } = config.params;
-    return new Promise((resolve, reject) => {
-      if (userMap[token]) {
-        setTimeout(() => {
-          resolve([200, {
-            data: userMap[token]
-          }]);
-        }, 100);
-      } else {
-        reject('获取失败')
-      }
-    })
+    const { token } = param2Obj(config.url);
+    if (userMap[token]) {
+      return userMap[token];
+    } else {
+      return Promise.reject('a');
+    }
   },
-  logout: () => new Promise(resolve => {
-    setTimeout(() => {
-      resolve([200, { data: 'success' }]);
-    }, 100);
-  })
+  logout: () => 'success'
 };
