@@ -1,4 +1,5 @@
 import Mock from 'mockjs';
+import { param2Obj } from 'utils';
 
 const NameList = [];
 const count = 100;
@@ -12,18 +13,12 @@ NameList.push({ name: 'mockPan' })
 
 export default {
   searchUser: config => {
-    const { name } = config.params;
+    const { name } = param2Obj(config.url);
     const mockNameList = NameList.filter(item => {
       const lowerCaseName = item.name.toLowerCase()
       if (name && lowerCaseName.indexOf(name.toLowerCase()) < 0) return false;
       return true;
     });
-    return new Promise(resolve => {
-      setTimeout(() => {
-        resolve([200, {
-          items: mockNameList
-        }]);
-      }, 100);
-    })
+    return { items: mockNameList }
   }
 };
