@@ -2,6 +2,7 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 import app from './modules/app';
 import user from './modules/user';
+import permission from './modules/permission';
 import getters from './getters';
 
 Vue.use(Vuex);
@@ -9,7 +10,8 @@ Vue.use(Vuex);
 const store = new Vuex.Store({
   modules: {
     app,
-    user
+    user,
+    permission
   },
   getters
 });
