@@ -5,7 +5,7 @@
 /* 是否是公司邮箱*/
 export function isWscnEmail(str) {
   const reg = /^[a-z0-9](?:[-_.+]?[a-z0-9]+)*@wallstreetcn\.com$/i;
-  return reg.test(str);
+  return reg.test(str.trim());
 }
 
 /* 合法uri*/
