@@ -1,14 +1,13 @@
 <template>
   <div class="components-container" style='height:100vh'>
     <div class='chart-container'>
-       <keyboardChart2 id='apple' height='100%' width='100%' />
+      <keyboardChart2 id='apple' height='100%' width='100%' />
     </div>
-
   </div>
 </template>
+
 <script>
     import keyboardChart2 from 'components/Charts/keyboard2';
-
     export default {
       components: { keyboardChart2 }
     };
@@ -16,9 +15,9 @@
 
 <style scoped>
 .chart-container{
-    position: relative;
-    width: 100%;
-    height: 90%;
+  position: relative;
+  width: 100%;
+  height: 90%;
 }
 </style>
 
