@@ -1,14 +1,13 @@
 <template>
   <div class="components-container" style='height:100vh'>
     <div class='chart-container'>
-       <mixchart id='apple' height='100%' width='100%' />
+      <mixchart id='apple' height='100%' width='100%' />
     </div>
-
   </div>
 </template>
+
 <script>
     import mixchart from 'components/Charts/mixchart';
-
     export default {
       components: { mixchart }
     };
@@ -16,10 +15,10 @@
 
 <style scoped>
 .chart-container{
-    position: relative;
-    width: 100%;
-    height: 90%;
-    padding-bottom: 40px;
+  position: relative;
+  width: 100%;
+  height: 90%;
+  padding-bottom: 40px;
 }
 </style>
 
