@@ -1,6 +1,6 @@
 <template>
   <div class="components-container">
-    <code>拖拽https://github.com/SortableJS/Vue.Draggable 项目：Vue.Draggable</code>
+    <code>drag-list base on <a href="https://github.com/SortableJS/Vue.Draggable" target="_blank">Vue.Draggable</a></code>
     <div class="editor-container">
        <DndList :list1="list1" :list2="list2" list1Title="头条列表" list2Title="文章池" />
     </div>
