@@ -4,7 +4,7 @@
     <el-row>
       <el-col :span="12">
         <h1 class="text-jumbo text-ginormous">Oops!</h1>
-         gif来源<a href='https://zh.airbnb.com/' target='_blank'>airbnb</a> 页面
+        gif来源<a href='https://zh.airbnb.com/' target='_blank'>airbnb</a> 页面
         <h2>你没有权限去该页面</h2>
         <h6>如有不满请联系你领导</h6>
         <ul class="list-unstyled">
@@ -13,7 +13,7 @@
             <router-link to="/dashboard">回首页</router-link>
           </li>
           <li class="link-type"><a href="https://www.taobao.com/">随便看看</a></li>
-          <li><a @click="dialogVisible=true" href="#">点我看图</a></li>
+          <li><a @click.prevent="dialogVisible=true" href="#">点我看图</a></li>
         </ul>
       </el-col>
       <el-col :span="12">
@@ -21,14 +21,14 @@
       </el-col>
     </el-row>
 
-    <el-dialog title="随便看" v-model="dialogVisible" size="large">
+    <el-dialog title="随便看" :visible.sync="dialogVisible" size="large">
       <img class="pan-img" :src="ewizardClap">
     </el-dialog>
   </div>
 </template>
+
 <script>
     import errGif from 'assets/401.gif';
-
     export default {
       data() {
         return {
@@ -39,45 +39,49 @@
       },
       methods: {
         back() {
-          this.$router.go(-1)
+          if (this.$route.query.noGoBack) {
+            this.$router.push({ path: '/dashboard' });
+          } else {
+            this.$router.go(-1);
+          }
         }
       }
     };
 </script>
 
 <style rel="stylesheet/scss" lang="scss" scoped>
-    .errPage-container {
-        width: 800px;
-        margin: 100px auto;
-        .pan-back-btn {
-            background: #008489;
-            color: #fff;
-        }
-        .pan-gif {
-            margin: 0 auto;
-            display: block;
-        }
-        .pan-img{
-            display: block;
-            margin: 0 auto;
-        }
-        .text-jumbo {
-            font-size: 60px;
-            font-weight: 700;
-            color: #484848;
-        }
-        .list-unstyled {
-            font-size: 14px;
-            li {
-                padding-bottom: 5px;
-            }
-            a {
-                color: #008489;
-                text-decoration: none;
-                &:hover {
-                    text-decoration: underline;
-                }
-            }
-        }
+  .errPage-container {
+    width: 800px;
+    margin: 100px auto;
+    .pan-back-btn {
+      background: #008489;
+      color: #fff;
     }
+    .pan-gif {
+      margin: 0 auto;
+      display: block;
+    }
+    .pan-img {
+      display: block;
+      margin: 0 auto;
+    }
+    .text-jumbo {
+      font-size: 60px;
+      font-weight: 700;
+      color: #484848;
+    }
+    .list-unstyled {
+      font-size: 14px;
+      li {
+        padding-bottom: 5px;
+      }
+      a {
+        color: #008489;
+        text-decoration: none;
+        &:hover {
+          text-decoration: underline;
+        }
+      }
+    }
+  }
 </style>
