@@ -9,7 +9,7 @@
       </div>
       <div class="bullshit">
         <div class="bullshit__oops">OOPS!</div>
-         <div class="bullshit__info">版权所有<a class='link-type' href='https://wallstreetcn.com' target='_blank'>华尔街见闻</a></div>
+        <div class="bullshit__info">版权所有<a class='link-type' href='https://wallstreetcn.com' target='_blank'>华尔街见闻</a></div>
         <div class="bullshit__headline">{{ message }}</div>
         <div class="bullshit__info">请检查您输入的网址是否正确，请点击以下按钮返回主页或者发送错误报告</div>
         <a href="/" class="bullshit__return-home">返回首页</a>
@@ -17,6 +17,7 @@
     </div>
   </div>
 </template>
+
 <script>
   export default {
     computed: {
@@ -26,6 +27,7 @@
     }
   }
 </script>
+
 <style rel="stylesheet/scss" lang="scss" scoped>
   .wscn-http404 {
     position: relative;
