@@ -11,9 +11,7 @@
       name: 'AppMain',
       computed: {
         key() {
-          return this.$route.name !== undefined
-                        ? this.$route.name + +new Date()
-                        : this.$route + +new Date()
+          return this.$route.name !== undefined ? this.$route.name + +new Date() : this.$route + +new Date()
         }
       }
     }
