@@ -1,24 +1,26 @@
 <template>
     <el-menu class="navbar" mode="horizontal">
-        <Hamburger class="hamburger-container" :toggleClick="toggleSideBar" :isActive="sidebar.opened"></Hamburger>
+        <hamburger class="hamburger-container" :toggleClick="toggleSideBar" :isActive="sidebar.opened"></hamburger>
         <levelbar></levelbar>
-        <ErrLog v-if="log.length>0" class="errLog-container" :logsList="log"></ErrLog>
+        <tabs-view></tabs-view>
+        <error-log v-if="log.length>0" class="errLog-container" :logsList="log"></error-log>
+        <screenfull class='screenfull'></screenfull>
         <el-dropdown class="avatar-container" trigger="click">
             <div class="avatar-wrapper">
                 <img class="user-avatar" :src="avatar+'?imageView2/1/w/80/h/80'">
-                <i class="el-icon-caret-bottom"/>
+                <i class="el-icon-caret-bottom"></i>
             </div>
             <el-dropdown-menu class="user-dropdown" slot="dropdown">
-                <router-link  class='inlineBlock' to="/">
+                <router-link class='inlineBlock' to="/">
                     <el-dropdown-item>
                         首页
                     </el-dropdown-item>
                 </router-link>
-                <router-link  class='inlineBlock' to="/admin/profile">
+                <a target='_blank' href="https://github.com/PanJiaChen/vue-element-admin/">
                     <el-dropdown-item>
-                        设置
+                        项目地址
                     </el-dropdown-item>
-                </router-link>
+                </a>
                 <el-dropdown-item divided><span @click="logout" style="display:block;">退出登录</span></el-dropdown-item>
             </el-dropdown-menu>
         </el-dropdown>
@@ -26,17 +28,21 @@
 </template>
 
 <script>
-    import { mapGetters } from 'vuex'
+    import { mapGetters } from 'vuex';
     import Levelbar from './Levelbar';
+    import TabsView from './TabsView';
     import Hamburger from 'components/Hamburger';
-    import ErrLog from 'components/ErrLog';
+    import Screenfull from 'components/Screenfull';
+    import ErrorLog from 'components/ErrLog';
     import errLogStore from 'store/errLog';
 
     export default {
       components: {
         Levelbar,
+        TabsView,
         Hamburger,
-        ErrLog
+        ErrorLog,
+        Screenfull
       },
       data() {
         return {
@@ -56,12 +62,13 @@
         },
         logout() {
           this.$store.dispatch('LogOut').then(() => {
-            this.$router.push({ path: '/login' })
+            location.reload();// 为了重新实例化vue-router对象 避免bug
           });
         }
       }
     }
 </script>
+
 <style rel="stylesheet/scss" lang="scss" scoped>
     .navbar {
         height: 50px;
@@ -77,6 +84,12 @@
             display: inline-block;
             position: absolute;
             right: 150px;
+        }
+        .screenfull{
+             position: absolute;
+             right: 90px;
+             top: 16px;
+             color: red;
         }
         .avatar-container {
             height: 50px;
