@@ -2,6 +2,6 @@ export { default as Navbar } from './Navbar';
 
 export { default as Sidebar } from './Sidebar';
 
-export { default as Levelbar } from './Sidebar';
+export { default as Levelbar } from './Levelbar';
 
 export { default as AppMain } from './AppMain';
