@@ -19,13 +19,13 @@
                     登录
                 </el-button>
             </el-form-item>
-            <div class='tips'>admin账号为:admin@wallstreetcn 密码随便填</div>
-            <div class='tips'>editor账号:editor@wallstreetcn 密码随便填</div>
+            <div class='tips'>admin账号为:admin@wallstreetcn.com 密码随便填</div>
+            <div class='tips'>editor账号:editor@wallstreetcn.com 密码随便填</div>
             <router-link to="/sendpwd" class="forget-pwd">
                 忘记密码?(或首次登录)
             </router-link>
         </el-form>
-        <el-dialog title="第三方验证" v-model="showDialog">
+        <el-dialog title="第三方验证" :visible.sync="showDialog">
             邮箱登录成功,请选择第三方验证
             <socialSign></socialSign>
         </el-dialog>
@@ -35,7 +35,7 @@
 <script>
     import { mapGetters } from 'vuex';
     import { isWscnEmail } from 'utils/validate';
-    import { getQueryObject } from 'utils';
+    // import { getQueryObject } from 'utils';
     import socialSign from './socialsignin';
 
     export default {
@@ -125,7 +125,7 @@
     }
 </script>
 
-<style rel="stylesheet/scss" lang="scss" scoprd>
+<style rel="stylesheet/scss" lang="scss">
     @import "src/styles/mixin.scss";
     .tips{
       font-size: 14px;
