@@ -19,15 +19,15 @@
       methods: {
         wechatHandleClick(thirdpart) {
           this.$store.commit('SET_AUTH_TYPE', thirdpart);
-          const appid = 'wxff5aaaad72308d46';
-          const redirect_uri = encodeURIComponent('http://wallstreetcn.com/auth/redirect?redirect=' + window.location.origin + '/authredirect');
+          const appid = 'xxxxx';
+          const redirect_uri = encodeURIComponent('xxx/redirect?redirect=' + window.location.origin + '/authredirect');
           const url = 'https://open.weixin.qq.com/connect/qrconnect?appid=' + appid + '&redirect_uri=' + redirect_uri + '&response_type=code&scope=snsapi_login#wechat_redirect';
           openWindow(url, thirdpart, 540, 540);
         },
         tencentHandleClick(thirdpart) {
           this.$store.commit('SET_AUTH_TYPE', thirdpart);
-          const client_id = '101150108';
-          const redirect_uri = encodeURIComponent('http://wallstreetcn.com/auth/redirect?redirect=' + window.location.origin + '/authredirect');
+          const client_id = 'xxxxx';
+          const redirect_uri = encodeURIComponent('xxx/redirect?redirect=' + window.location.origin + '/authredirect');
           const url = 'https://graph.qq.com/oauth2.0/authorize?response_type=code&client_id=' + client_id + '&redirect_uri=' + redirect_uri;
           openWindow(url, thirdpart, 540, 540);
         }
