@@ -1,10 +1,10 @@
 <template>
   <div class="app-container">
-     <div style='margin-bottom:15px;'>你的权限： {{roles}}</div>
-     切换权限：
-      <el-radio-group v-model="role">
-        <el-radio-button label="editor"></el-radio-button>
-      </el-radio-group>
+    <div style='margin-bottom:15px;'>你的权限： {{roles}}</div>
+    切换权限：
+    <el-radio-group v-model="role">
+      <el-radio-button label="editor"></el-radio-button>
+    </el-radio-group>
   </div>
 </template>
 
@@ -23,8 +23,9 @@
       },
       watch: {
         role(val) {
-          this.$store.commit('SET_ROLES', [val]);
-          window.location.reload()
+          this.$store.dispatch('ChangeRole', val).then(() => {
+            this.$router.push({ path: '/permission/index?' + +new Date() });
+          })
         }
       }
     }
